fix(list): Fix segmentation fault due to missing error handling
The stub implementation of the ListDrives function returns a
`response, error` tuple containing either the list of drives, or an
error message indicating that the function is not implemented on the
current platform.

The calling code does not verify that if the function returns an error
or a response, and attempts to iterate through the null response
causing a segmentation fault on Linux.

To fix, check that the response is valid before accessing it and wrap
and return the error if the ListDrives function generated one.